Componentize product details modal
