+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/projects/[id=idNumber]/+page.svelte
@@ -10,6 +10,7 @@
   import { ProductTransitionType, RoleId } from 'sil.appbuilder.portal.common/prisma';
   import { superForm } from 'sveltekit-superforms';
   import type { PageData } from './$types';
+  import ProductDetails from '$lib/components/ProductDetails.svelte';
 
   const langtagmap = new Map(langtags.map((tag) => [tag.tag, /* tag.localname ?? */ tag.name]));
 
@@ -37,28 +38,6 @@
   );
   function openModal(id: string) {
     (window[('modal' + id) as any] as any).showModal();
-  }
-  function stateString(workflowTypeNum: number, transitionType: number) {
-    const workflowType = (
-      m[
-        ('admin_settings_workflowDefinitions_workflowTypes_' + workflowTypeNum) as keyof typeof m
-      ] as any
-    )();
-    switch (transitionType) {
-      case 2:
-        return m.project_products_transitions_transitionTypes_2({
-          workflowType
-        });
-      case 3:
-        return m.project_products_transitions_transitionTypes_3({
-          workflowType
-        });
-      case 4:
-        return m.project_products_transitions_transitionTypes_4({
-          workflowType
-        });
-    }
-    return '';
   }
   let simpleSettingsFormTimeout: NodeJS.Timeout;
   let simpleSettingsForm: HTMLFormElement;
@@ -230,84 +209,7 @@
                   </a>
                 {/if}
               </div>
-              <dialog id="modal{product.Id}" class="modal">
-                <div class="modal-box w-11/12 max-w-6xl">
-                  <h2>{m.project_products_transitions_productDetails()}</h2>
-                  <table class="table">
-                    <thead>
-                      <tr>
-                        <th>{m.project_products_transitions_storeName()}</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      <tr>
-                        <td>
-                          {product.Store?.Description}
-                        </td>
-                      </tr>
-                    </tbody>
-                  </table>
-                  <table class="table">
-                    <thead>
-                      <tr>
-                        <th>{m.project_products_transitions_state()}</th>
-                        <th>{m.project_products_transitions_user()}</th>
-                        <th>{m.project_products_transitions_command()}</th>
-                        <th>{m.project_products_transitions_comment()}</th>
-                        <th>{m.project_products_transitions_date()}</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {#each product.Transitions as transition}
-                        <tr class:font-bold={[2, 3, 4].includes(transition.TransitionType)}>
-                          <td>
-                            {#if transition.TransitionType === ProductTransitionType.Activity}
-                              {transition.InitialState}
-                            {:else if transition.TransitionType === ProductTransitionType.ProjectAccess}
-                              * {transition.InitialState}
-                            {:else}
-                              {stateString(transition.WorkflowType ?? 1, transition.TransitionType)}
-                            {/if}
-                          </td>
-                          <td>
-                            <!-- Does not include WorkflowUserId mapping. Might be needed but didn't seem like it to me -->
-                            {#if ![2, 3, 4].includes(transition.TransitionType)}
-                              {transition.AllowedUserNames || m.appName()}
-                            {/if}
-                            <!-- TODO: handle username for Project Access transition types -->
-                          </td>
-                          <td>{transition.Command ?? ''}</td>
-                          <td>
-                            {#if transition.Comment?.startsWith('system.')}
-                              {#if transition.Comment.startsWith('system.build-failed')}
-                                <span>
-                                  {m.system_buildFailed()}
-                                </span>
-                              {:else if transition.Comment.startsWith('system.publish-failed')}
-                                <span>
-                                  {m.system_publishFailed()}
-                                </span>
-                              {/if}
-                              <br />
-                              <a href={transition.Comment.replace('system.build-failed,', '')}>
-                                {m.project_products_publications_console()}
-                              </a>
-                            {:else}
-                              {transition.Comment ?? ''}
-                            {/if}
-                          </td>
-                          <td>
-                            {getTimeDateString(transition.DateTransition)}
-                          </td>
-                        </tr>
-                      {/each}
-                    </tbody>
-                  </table>
-                </div>
-                <form method="dialog" class="modal-backdrop">
-                  <button>{m.project_side_authors_close()}</button>
-                </form>
-              </dialog>
+              <ProductDetails product={product} />
             </div>
           {/each}
         {/if}
